[PDS-167] refactor: 장비 및 차량 공통 dto 위치 변경
