+++ src/main/java/com/example/pladialmserver/booking/service/ProductBookingService.java
@@ -3,7 +3,7 @@ package com.example.pladialmserver.booking.service;
 import com.example.pladialmserver.booking.dto.request.ReturnProductReq;
 import com.example.pladialmserver.booking.dto.response.BookingRes;
 import com.example.pladialmserver.booking.dto.response.ProductBookingDetailRes;
-import com.example.pladialmserver.product.resource.dto.response.AdminProductRes;
+import com.example.pladialmserver.product.dto.response.AdminProductRes;
 import com.example.pladialmserver.user.entity.User;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
