+++ src/main/java/com/example/pladialmserver/office/dto/request/CreateOfficeReq.java
@@ -1,4 +1,4 @@
-package com.example.pladialmserver.product.resource.dto.request;
+package com.example.pladialmserver.office.dto.request;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Data;
