+++ src/main/java/com/example/pladialmserver/product/car/controller/CarAdminController.java
@@ -3,9 +3,9 @@ package com.example.pladialmserver.product.car.controller;
 import com.example.pladialmserver.global.resolver.Account;
 import com.example.pladialmserver.global.response.ResponseCustom;
 import com.example.pladialmserver.product.car.service.CarService;
-import com.example.pladialmserver.product.resource.dto.request.CreateProductReq;
-import com.example.pladialmserver.product.resource.dto.response.AdminProductDetailsRes;
-import com.example.pladialmserver.product.resource.dto.response.AdminProductsRes;
+import com.example.pladialmserver.product.dto.request.CreateProductReq;
+import com.example.pladialmserver.product.dto.response.AdminProductDetailsRes;
+import com.example.pladialmserver.product.dto.response.AdminProductsRes;
 import com.example.pladialmserver.user.entity.User;
 import io.swagger.annotations.Api;
 import io.swagger.v3.oas.annotations.Operation;
