+++ src/main/java/com/example/pladialmserver/product/car/service/CarService.java
@@ -16,10 +16,10 @@ import com.example.pladialmserver.product.car.repository.CarRepository;
 import com.example.pladialmserver.product.dto.request.ProductReq;
 import com.example.pladialmserver.product.dto.response.ProductBookingRes;
 import com.example.pladialmserver.product.dto.response.ProductDetailRes;
-import com.example.pladialmserver.product.resource.dto.request.CreateProductReq;
-import com.example.pladialmserver.product.resource.dto.response.AdminProductDetailsRes;
-import com.example.pladialmserver.product.resource.dto.response.AdminProductsRes;
-import com.example.pladialmserver.product.resource.dto.response.ProductList;
+import com.example.pladialmserver.product.dto.request.CreateProductReq;
+import com.example.pladialmserver.product.dto.response.AdminProductDetailsRes;
+import com.example.pladialmserver.product.dto.response.AdminProductsRes;
+import com.example.pladialmserver.product.dto.response.ProductList;
 import com.example.pladialmserver.product.service.ProductService;
 import com.example.pladialmserver.user.entity.Role;
 import com.example.pladialmserver.user.entity.User;
