+++ src/main/java/com/example/pladialmserver/product/resource/repository/ResourceCustom.java
@@ -1,7 +1,7 @@
 package com.example.pladialmserver.product.resource.repository;
 
-import com.example.pladialmserver.product.resource.dto.response.AdminProductsRes;
-import com.example.pladialmserver.product.resource.dto.response.ResourceRes;
+import com.example.pladialmserver.product.dto.response.AdminProductsRes;
+import com.example.pladialmserver.product.resource.dto.ResourceRes;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
