+++ src/main/java/com/example/pladialmserver/product/resource/repository/ResourceRepositoryImpl.java
@@ -1,9 +1,9 @@
 package com.example.pladialmserver.product.resource.repository;
 
-import com.example.pladialmserver.product.resource.dto.response.AdminProductsRes;
-import com.example.pladialmserver.product.resource.dto.response.QAdminResourcesRes;
-import com.example.pladialmserver.product.resource.dto.response.QResourceRes;
-import com.example.pladialmserver.product.resource.dto.response.ResourceRes;
+import com.example.pladialmserver.product.dto.response.AdminProductsRes;
+import com.example.pladialmserver.product.dto.response.QAdminProductsRes;
+import com.example.pladialmserver.product.resource.dto.QResourceRes;
+import com.example.pladialmserver.product.resource.dto.ResourceRes;
 import com.querydsl.core.types.Predicate;
 import com.querydsl.core.types.dsl.BooleanExpression;
 import com.querydsl.jpa.impl.JPAQueryFactory;
@@ -30,7 +30,7 @@ public class ResourceRepositoryImpl implements ResourceCustom {
     @Override
     public Page<AdminProductsRes> search(String resourceName, Pageable pageable) {
         List<AdminProductsRes> results = jpaQueryFactory
-                .select(new QAdminResourcesRes(
+                .select(new QAdminProductsRes(
                         resource.resourceId,
                         resource.name,
                         resource.manufacturer,
