+++ src/main/java/com/example/pladialmserver/product/resource/service/ResourceService.java
@@ -12,11 +12,11 @@ import com.example.pladialmserver.global.utils.EmailUtil;
 import com.example.pladialmserver.product.dto.request.ProductReq;
 import com.example.pladialmserver.product.dto.response.ProductBookingRes;
 import com.example.pladialmserver.product.dto.response.ProductDetailRes;
-import com.example.pladialmserver.product.resource.dto.request.CreateProductReq;
-import com.example.pladialmserver.product.resource.dto.response.AdminProductDetailsRes;
-import com.example.pladialmserver.product.resource.dto.response.AdminProductsRes;
-import com.example.pladialmserver.product.resource.dto.response.ProductList;
-import com.example.pladialmserver.product.resource.dto.response.ResourceRes;
+import com.example.pladialmserver.product.dto.request.CreateProductReq;
+import com.example.pladialmserver.product.dto.response.AdminProductDetailsRes;
+import com.example.pladialmserver.product.dto.response.AdminProductsRes;
+import com.example.pladialmserver.product.dto.response.ProductList;
+import com.example.pladialmserver.product.resource.dto.ResourceRes;
 import com.example.pladialmserver.product.resource.entity.Resource;
 import com.example.pladialmserver.product.resource.repository.ResourceRepository;
 import com.example.pladialmserver.product.service.ProductService;
